add references and finish structure
