+++ Chirp.sln
@@ -7,8 +7,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{827E0CD3-B72
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Web", "src\Chirp.Web\Chirp.Web.csproj", "{8991C2A3-4245-4557-A531-7BEDD16A5330}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.SQLite", "src\Chirp.SQLite\Chirp.SQLite.csproj", "{C0C34AAC-E070-4153-9BF7-B5126879B1ED}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{0C88DD14-F956-CE84-757C-A364CCF449FC}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Client_Test", "test\Client_Test\Client_Test.csproj", "{B8A7204A-2F43-4A78-A877-0B8BEB817407}"
@@ -41,18 +39,6 @@ Global
 		{8991C2A3-4245-4557-A531-7BEDD16A5330}.Release|x64.Build.0 = Release|Any CPU
 		{8991C2A3-4245-4557-A531-7BEDD16A5330}.Release|x86.ActiveCfg = Release|Any CPU
 		{8991C2A3-4245-4557-A531-7BEDD16A5330}.Release|x86.Build.0 = Release|Any CPU
-		{C0C34AAC-E070-4153-9BF7-B5126879B1ED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C0C34AAC-E070-4153-9BF7-B5126879B1ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C0C34AAC-E070-4153-9BF7-B5126879B1ED}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{C0C34AAC-E070-4153-9BF7-B5126879B1ED}.Debug|x64.Build.0 = Debug|Any CPU
-		{C0C34AAC-E070-4153-9BF7-B5126879B1ED}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{C0C34AAC-E070-4153-9BF7-B5126879B1ED}.Debug|x86.Build.0 = Debug|Any CPU
-		{C0C34AAC-E070-4153-9BF7-B5126879B1ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C0C34AAC-E070-4153-9BF7-B5126879B1ED}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C0C34AAC-E070-4153-9BF7-B5126879B1ED}.Release|x64.ActiveCfg = Release|Any CPU
-		{C0C34AAC-E070-4153-9BF7-B5126879B1ED}.Release|x64.Build.0 = Release|Any CPU
-		{C0C34AAC-E070-4153-9BF7-B5126879B1ED}.Release|x86.ActiveCfg = Release|Any CPU
-		{C0C34AAC-E070-4153-9BF7-B5126879B1ED}.Release|x86.Build.0 = Release|Any CPU
 		{B8A7204A-2F43-4A78-A877-0B8BEB817407}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B8A7204A-2F43-4A78-A877-0B8BEB817407}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B8A7204A-2F43-4A78-A877-0B8BEB817407}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -107,7 +93,6 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{8991C2A3-4245-4557-A531-7BEDD16A5330} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
-		{C0C34AAC-E070-4153-9BF7-B5126879B1ED} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{B8A7204A-2F43-4A78-A877-0B8BEB817407} = {0C88DD14-F956-CE84-757C-A364CCF449FC}
 		{CF948033-CAEA-4185-895C-BA845A75DF10} = {0C88DD14-F956-CE84-757C-A364CCF449FC}
 		{14A76D42-0A1C-4010-8FA0-2E8AFF2D52C3} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
