+++ src/Chirp.Infrastructure/AuthorDTO.cs
@@ -1,0 +1,6 @@
+namespace Chirp.Infrastructure;
+
+public class AuthorDTO
+{
+    
+}
