+++ src/Chirp.Infrastructure/CheepDBContext.cs
@@ -1,6 +1,7 @@
+using Chirp.Core;
 using Microsoft.EntityFrameworkCore;
 
-namespace Chirp.Razor;
+namespace Chirp.Infrastructure;
 
 
 public class CheepDBContext : DbContext
