+++ src/Chirp.Infrastructure/CheepDTO.cs
@@ -1,0 +1,6 @@
+namespace Chirp.Infrastructure;
+
+public class CheepDTO
+{
+    
+}
