+++ src/Chirp.Infrastructure/CheepRepository.cs
@@ -1,8 +1,7 @@
 namespace Chirp.Infrastructure;
 
-public interface ICheepRepository
+
+public class CheepRepository : ICheepRepository
 {
-}
-public class CheepRepository
-{
+    
 }
