+++ src/Chirp.Infrastructure/CheepService.cs
@@ -2,15 +2,6 @@ using Chirp.Infrastructure;
 //using Microsoft.Data.Sqlite;
 public record CheepViewModel(string Author, string Message, string Timestamp);
 
-public interface ICheepService
-{
-    public List<CheepViewModel> GetCheeps();
-    public List<CheepViewModel> GetCheeps(int page, int pageSize);
-    public List<CheepViewModel> GetCheepsFromAuthor(string author);
-
-    public List<CheepViewModel> GetCheepsFromAuthor(string author, int page, int pageSize);
-}
-
 public class CheepService : ICheepService
 {
     public List<CheepViewModel> GetCheeps()
