+++ src/Chirp.Infrastructure/Interfaces/ICheepRepository.cs
@@ -1,0 +1,6 @@
+namespace Chirp.Infrastructure;
+
+public interface ICheepRepository
+{
+    
+}
