--- src/Chirp.SQLite/DBinit.sh
@@ -1,3 +1,0 @@
-# Creates the database locally in temp
-sqlite3 /tmp/chirp.db < data/schema.sql
-sqlite3 /tmp/chirp.db < data/dump.sql
