--- src/Chirp.SQLite/scripts/initDB.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-sqlite3 /tmp/chirp.db < data/schema.sql
-sqlite3 /tmp/chirp.db < data/dump.sql
