--- src/Chirp.Web/Author.cs
@@ -1,21 +1,0 @@
-namespace Chirp.Razor;
-using System.ComponentModel.DataAnnotations;
-
-public class Author
-{
-    [Required]
-    public required int AuthorID  { get; set; }
-    
-    [Required]
-    public required string name { get; set; }
-    
-    [Required]
-    public required string email { get; set; }
-    
-    
-    public ICollection<Cheep>  cheeps { get; set; }
-
-   
-    
-    
-}
